requirements.txt
python-docx==1.1.2
PyMuPDF==1.24.9
moviepy==1.0.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
numpy>=1.24,<3
python-dotenv==1.0.1
requests>=2.32.0
google-genai>=0.3.0


